refactor!: Start from a clean slate
The flake-parts stuff was just adding accidental complexity. This
whole repository basically only provides a single, stupid simple
overlay that contains one function, and a home-manager
module. `perSystem` isn't even used, except for an example.

Rather than maintaining all of that, and adding a workflow to update
the flake inputs, let's just simplify it down to an overlay, and
document how to use that overlay in dependant flakes.

If we still want to add example packages later we can of course do so,
but I consider that overkill for now.